basic resource placement algorithm
This algorithm was tested only in debug local mode.
Alogorithm seems to work, and distributes resources on nodes with enough
resources.

This was tested with b.basic_placeResources() on line brain.go:389. It
didn't need master node for testing. In this commit I move that function
call into master only part of this for loop. (brain.go:393)